glispのdraw.rsを分割
